+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * #117 Fix table row loop missing ``key`` prop
 * #122 Fix async tab-items
+* #124 Toast, Dialog, Snackbar and Programmatic Modal exposed
 
 ## 0.4.2